# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Suas Respostas",
        "description": " API para plataformas que visa permitir que pessoas publiquem questões e respostas de atividades e trabalhos.",
        "version": "1.0",
        "contact": {
            "email": "geoo677@gmail.com"
        }
    },
    "servers": [
        {
            "url": "https://api-suas-questoes.herokuapp.com/",
            "description": "API em produção",
            "security": [
                {
                    "bearerAuth": []
                }
            ]
        },
        {
            "url": "http://localhost:8080/",
            "description": "API localhost",
            "security": [
                {
                    "bearerAuth": []
                }
            ]
        }
    ],
    "basePath": "/",
    "paths": {
        "/questions?page={page}": {
            "get": {
                "tags": [
                    "Questões"
                ],
                "summary": "Lista questões",
                "description": "Retorna uma lista de questões.",
                "operationId": "getQuestoes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Página das questões.",
                        "required": false,
                        "default": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna lista de questões",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Questions"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao listar questões",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/questions/?statement={statement}&page={page}": {
            "get": {
                "summary": "Buscar questão por enunciado",
                "description": "Busca contenha enuciado informado pela query statement.",
                "operationId": "findQuestion",
                "tags": [
                    "Questões"
                ],
                "parameters": [
                    {
                        "name": "statement",
                        "in": "path",
                        "description": "Enuciado da questão",
                        "required": true,
                        "default": "água"
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "página",
                        "required": false,
                        "default": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna lista de vídeos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Questions"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar questão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/questions": {
            "post": {
                "summary": "Salvar questão",
                "description": "Salva questão via post.",
                "tags": [
                    "Questões"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "postQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Question"
                            },
                            "examples": {
                                "question": {
                                    "value": {
                                        "matter": "Matemática",
                                        "statement": "1+1",
                                        "answer": "=2 (Resposta não obrigatória)"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Confirmação de que a questão foi salvo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acesso negado!.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao salvar questão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/questions/{id}": {
            "get": {
                "tags": [
                    "Questões"
                ],
                "summary": "Obter questão.",
                "description": "Retornar a questão do id informado pelo parâmetro da URL.",
                "operationId": "getQuestoe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da questão que se deseja buscar.",
                        "required": true,
                        "default": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma questão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Questão não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar questão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Questões"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Editar questão.",
                "description": "Edita o questão do id informado pelo parâmetro  da URL.",
                "operationId": "putVideo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da questão que deseja editar.",
                        "required": true,
                        "default": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Question"
                            },
                            "examples": {
                                "question": {
                                    "value": {
                                        "matter": "Matemática",
                                        "statement": "1+1",
                                        "answer": "=2 (Resposta não obrigatória)"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Confirmação de que a questão foi editado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acesso negado!.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Questão não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao editar questão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Questões"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Deletar questão.",
                "description": "Deleta a questão do id informado pelo parâmetro da URL.",
                "operationId": "deleteQuestão",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da questão que deseja deletar.",
                        "required": true,
                        "default": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmação de que a questão deletada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acesso negado!.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Questão não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao deletar questão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/questions/by/user": {
            "get": {
                "tags": [
                    "Questões"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Obter questão do usuário.",
                "description": "Retornar as questões postadas pelo usuário.",
                "operationId": "getQuestoeByUser",
        
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de questões.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Questions"
                                }
                            }
                        }
                    },
                   
                    "500": {
                        "description": "Erro ao buscar questões",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Message": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "Question": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "matter": {
                        "type": "string"
                    },
                    "statement": {
                        "type": "string"
                    },
                    "answer": {
                        "type": "string"
                    },
                    "user_name": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                }
            },
...
            "Questions": {
                "type": "object",
                "properties": {
                    "limit": {
                        "type": "integer"
                    },
                    "page": {
                        "type": "integer"
                    },
                    "count": {
                        "type": "integer"
                    },
                    "offset": {
                        "type": "integer"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}